Compression should now be working for fluorescence images (only tested for pharyngeal markers). Run it by calling compressMultipleFilesFun ... --pattern_include 'nameOfFirstImage.dat' --json_file fluorescence/pharynx.json

+++ fluorescence/readDat.py
@@ -43,7 +43,7 @@ class readDat:
             image_decoded = np.zeros((self.height, self.width), np.uint16)
             image_decoded[::-1, -2::-2] = D3s.reshape((self.height, -1))
             image_decoded[::-1, ::-2] = D1s.reshape((self.height, -1))
-            return (1, image_decoded)
+            return (1, image_decoded.astype(np.uint8))
         else:
             return (0, [])
     def release(self):
